readCRU
I made an R script similar to that of read Crisis.
Since the dataset is small, I made a while loop to do the geocode()
repeatedly until the NA for lon and lat is gone.

I made sure there is a maximum reptition which prevents the while loop
to run infinitely. By doing this, I could have managed to have
significantly less NA values in coords. (in fact I have all of them)